[#29] refactor : batch 모듈 ktlint 적용
