+++ coupon-platform-batch/src/main/kotlin/coupon_platform/batch/CouponPlatformBatchApplication.kt
