+++ coupon-platform-batch/src/main/kotlin/coupon_platform/batch/attendancecheck/job/ThirtyDayAttendanceCheckJobConfig.kt
@@ -57,5 +57,4 @@ class ThirtyDayAttendanceCheckJobConfig(
         .tasklet(thirtyDayAttendanceCheckTasklet, platformTransactionManager)
         .listener(stepListener)
         .build()
-
 }
