+++ coupon-platform-batch/src/main/kotlin/coupon_platform/batch/attendancecheck/job/ThreeDayAttendanceCheckJobConfig.kt
@@ -57,5 +57,4 @@ class ThreeDayAttendanceCheckJobConfig(
         .tasklet(threeDayAttendanceCheckTasklet, platformTransactionManager)
         .listener(stepListener)
         .build()
-
 }
