+++ coupon-platform-batch/src/main/kotlin/coupon_platform/batch/issuedcoupon/IssueCouponToRedisTasklet.kt
@@ -1,9 +1,9 @@
-package coupon_platform.batch.issued_coupon
+package coupon_platform.batch.issuedcoupon
 
 import coupon_platform.domain.issuedcoupon.entitiy.IssuedCoupon
 import coupon_platform.domain.issuedcoupon.repository.IssuedCouponReader
-import coupon_platform.infrastructure.cache.global_redis.handler.RedisHandlerOfIssuedCoupon
-import coupon_platform.infrastructure.cache.global_redis.util.CacheConstants.ISSUED_COUPON_KEY_TTL
+import coupon_platform.infrastructure.cache.globalredis.handler.RedisHandlerOfIssuedCoupon
+import coupon_platform.infrastructure.cache.globalredis.util.CacheConstants.ISSUED_COUPON_KEY_TTL
 import org.springframework.batch.core.StepContribution
 import org.springframework.batch.core.scope.context.ChunkContext
 import org.springframework.batch.core.step.tasklet.Tasklet
@@ -15,7 +15,7 @@ import java.time.LocalDate
 class IssueCouponToRedisTasklet(
     private val issuedCouponReader: IssuedCouponReader,
     private val redisHandler: RedisHandlerOfIssuedCoupon,
-): Tasklet {
+) : Tasklet {
 
     override fun execute(contribution: StepContribution, chunkContext: ChunkContext): RepeatStatus? {
         val findIssuedCoupons: List<IssuedCoupon> = findIssuedCoupons()
@@ -37,4 +37,3 @@ class IssueCouponToRedisTasklet(
         redisHandler.setIssuedCouponsWithTTL(findIssuedCoupons, ISSUED_COUPON_KEY_TTL)
     }
 }
-
