+++ coupon-platform-batch/src/main/kotlin/coupon_platform/batch/issuedcoupon/SaveIssuedCouponToRedisJob.kt
@@ -1,6 +1,6 @@
-package coupon_platform.batch.issued_coupon
+package coupon_platform.batch.issuedcoupon
 
-import coupon_platform.batch.issued_coupon.SaveIssuedCouponToRedisJob.Companion.JOB_NAME_OF_ISSUE_COUPON
+import coupon_platform.batch.issuedcoupon.SaveIssuedCouponToRedisJob.Companion.JOB_NAME_OF_ISSUE_COUPON
 import coupon_platform.batch.listener.JobListener
 import coupon_platform.batch.listener.StepListener
 import org.springframework.batch.core.Job
