+++ coupon-platform-batch/src/main/kotlin/coupon_platform/batch/listener/StepListener.kt
@@ -28,5 +28,4 @@ class StepListener {
 
         return exitStatus
     }
-
 }
